Refactor test and production code
