+++ HackrNewsApp/HackrNewsApp/App/SceneDelegate.swift
@@ -11,7 +11,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     var window: UIWindow?
     private let baseUrl = Endpoint.baseUrl
     var previousViewController: UIViewController?
-    var tabBarAction: ((UIViewController) -> Void)?
+    var tabBarAction: ((HackrNewsFeedViewController) -> Void)? = { $0.scrollToTop() }
 
     private lazy var httpClient: HTTPClient = {
         URLSessionHTTPClient(session: URLSession(configuration: .ephemeral))
@@ -39,7 +39,6 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         guard let scene = (scene as? UIWindowScene) else { return }
         window = UIWindow(windowScene: scene)
         configureWindow()
-        tabBarAction = tabBarAction(to:)
     }
 
     func configureWindow() {
@@ -112,18 +111,14 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
             )
         )
     }
-
-    private func tabBarAction(to viewController: UIViewController) {
-        if let viewController = (viewController as? UINavigationController)?.topViewController as? HackrNewsFeedViewController {
-            viewController.scrollToTop()
-        }
-    }
 }
 
 extension SceneDelegate: UITabBarControllerDelegate {
     func tabBarController(_: UITabBarController, didSelect viewController: UIViewController) {
         if previousViewController == viewController {
-            tabBarAction?(viewController)
+            if let viewController = (viewController as? UINavigationController)?.topViewController as? HackrNewsFeedViewController {
+                tabBarAction?(viewController)
+            }
         }
         previousViewController = viewController
     }
